Switch to floating point approach and fix at negative zero

+++ amd/device-libs/ocml/src/predH.cl
@@ -11,10 +11,10 @@ REQUIRES_16BIT_INSTS CONSTATTR half
 MATH_MANGLE(pred)(half x)
 {
     short ix = AS_SHORT(x);
-    short mx = SIGNBIT_HP16 - ix;
+    short mx = (short)SIGNBIT_HP16 - ix;
     mx = ix < (short)0 ? mx : ix;
-    short t = mx - BUILTIN_CLASS_F16(x, CLASS_NNOR|CLASS_NSUB|CLASS_NZER|CLASS_PZER|CLASS_PSUB|CLASS_PNOR|CLASS_PINF);
-    short r = SIGNBIT_HP16 - t;
+    short t = mx - (short)BUILTIN_CLASS_F16(x, CLASS_NNOR|CLASS_NSUB|CLASS_NZER|CLASS_PZER|CLASS_PSUB|CLASS_PNOR|CLASS_PINF);
+    short r = (short)SIGNBIT_HP16 - t;
     r = t < (short)0 ? r : t;
     return AS_HALF(r);
 }
